refactor(pointer): split button handling into frame, press, and release modules
Move pointer button handling into a folder-based module layout with
button/mod.rs as the coordinator entry point.

- add button/frame.rs for frame context, binding matching, and timing helpers
- add button/press.rs for press handlers and press-side helpers
- add button/release.rs for release cleanup and finalization paths
- add button/tests.rs for button-specific tests
- move button.rs to button/mod.rs

This improves cohesion, reduces per-file surface area, and makes
press/release behavior easier to change independently.

Verified with:
- cargo fmt --all
- cargo test -p halley-wl
- cargo check
